requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
msal==1.25.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-data-tables==12.4.4
azure-cosmos==4.5.1

# Frontend dependencies  
streamlit==1.28.2
requests==2.31.0

# Common dependencies
python-multipart==0.0.6
python-dotenv==1.0.0
